requirements.txt
torch>=2.2.0
transformers>=4.41.0
datasets>=2.19.0
sentencepiece>=0.1.99
protobuf>=4.25.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
pyarrow>=14.0.0
evaluate>=0.4.0
tqdm>=4.66.0

sentence-transformers>=2.6.0
spacy>=3.7.0

# python -m spacy download en_core_web_sm

fastopic>=0.1.6
topmost>=0.0.8

streamlit>=1.33.0
plotly>=5.20.0
matplotlib>=3.8.0
wordcloud>=1.9.3

python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0

praw>=7.7.1
prawcore>=2.4.0
google-genai>=0.3.0

